myers: move accept arg to kernel, introduce Chunk.details

+++ sdiff/cython/compare.pxd
@@ -1,2 +1,2 @@
 cdef class ComparisonBackend:
-    cdef double compare(self, Py_ssize_t i, Py_ssize_t j)
+    cdef int compare(self, Py_ssize_t i, Py_ssize_t j)
